requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pymongo==3.11.1
python-dotenv==0.15.0
Werkzeug==1.0.1

redis~=3.5.3